I2C: 16-bit register addresses + raw read/write
Widen register address from uint8_t to uint16_t across LL and HAL:
- If reg <= 0xFF: sends 1 address byte (backward compatible)
- If reg > 0xFF: sends 2 address bytes MSB-first

This supports devices like the DAC63202W (TI) that use wider
register addresses while keeping all existing 8-bit drivers
working unchanged.

Raw I2C read/write (no register byte) already existed in both
LL and HAL layers — no changes needed for command-based devices
like the STC31-C or MAX11644.

+++ sdk/hal/hal_i2c.c
@@ -79,25 +79,25 @@ hal_status_t hal_i2c_read(hal_i2c_t *h, uint8_t addr,
  * ============================================================ */
 
 hal_status_t hal_i2c_write_reg(hal_i2c_t *h, uint8_t addr,
-                               uint8_t reg, const uint8_t *data, uint32_t len)
+                               uint16_t reg, const uint8_t *data, uint32_t len)
 {
     return _convert_status(ll_i2c_write_reg(h->instance, addr, reg, data, len));
 }
 
 hal_status_t hal_i2c_read_reg(hal_i2c_t *h, uint8_t addr,
-                              uint8_t reg, uint8_t *buf, uint32_t len)
+                              uint16_t reg, uint8_t *buf, uint32_t len)
 {
     return _convert_status(ll_i2c_read_reg(h->instance, addr, reg, buf, len));
 }
 
 hal_status_t hal_i2c_write_byte(hal_i2c_t *h, uint8_t addr,
-                                uint8_t reg, uint8_t value)
+                                uint16_t reg, uint8_t value)
 {
     return hal_i2c_write_reg(h, addr, reg, &value, 1);
 }
 
 hal_status_t hal_i2c_read_byte(hal_i2c_t *h, uint8_t addr,
-                               uint8_t reg, uint8_t *value)
+                               uint16_t reg, uint8_t *value)
 {
     return hal_i2c_read_reg(h, addr, reg, value, 1);
 }
